Added comments. Added way for the main raindrop to be randomly selected.

+++ app/src/main/java/up/edu/raindrops/Drops.java
@@ -3,13 +3,19 @@ package up.edu.raindrops;
 import android.graphics.Color;
 import android.graphics.Paint;
 
+/**
+ * Custom class Drops that stores the information of a specific Raindrop
+ * Stores position, radius, and the color of the object
+ * @author Sean Yang
+ * @version 1.0 2-11-25
+ */
 public class Drops {
-    private int x;
-    private int y;
-    private int radius;
+    private float x;
+    private float y;
+    private float radius;
     private Paint color;
 
-    public Drops(int x, int y, int radius, Paint color) {
+    public Drops(float x, float y, float radius, Paint color) {
         this.x = x;
         this.y = y;
         this.radius = radius;
